auth för färdigställning

+++ migrations/00003_create_products_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE `products` (
   `product_id` int NOT NULL AUTO_INCREMENT,
   `product_name` varchar(25) NOT NULL UNIQUE KEY,
   `stock`int NOT NULL DEFAULT 0,
-  `price` float NOT NULL,
+  `price` DECIMAL(10,2) NOT NULL,
   `manufacturer` varchar(100) NOT NULL,
   `description`varchar(255) NOT NULL,
   `category_name` varchar(25),
